+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cliffs;
+pub mod contours;
+pub mod knolls;
 pub mod merge;
 pub mod util;
-pub mod contours;
